Changed the way one global variable is defined and updated the project file to include/exclude relevant files.
when we defined twi_interface in twi_interface.c we did it statically,
but in order to allow it to be global and used with the extern keyword
in other modules we can't use the static keywork defining it. We just
use twi_interface_t.

The project file needs to be updated to exclude the old driver as well
as automatically include the new files.